Fixes for re-enabling more MSVC level 4 warnings: cc/ edition
This contains fixes for the following sorts of issues:
* Signedness mismatch

BUG=81439
TEST=none

Review URL: https://codereview.chromium.org/363093005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@281205 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 07b48b0d253f9e6b2e5606cc53077f5971a88d64

+++ cc/resources/task_graph_runner.cc
@@ -17,7 +17,10 @@ namespace {
 class DependentIterator {
  public:
   DependentIterator(TaskGraph* graph, const Task* task)
-      : graph_(graph), task_(task), current_index_(-1), current_node_(NULL) {
+      : graph_(graph),
+        task_(task),
+        current_index_(static_cast<size_t>(-1)),
+        current_node_(NULL) {
     ++(*this);
   }
